Clean up query concerns
Move query-related concerns so they are all under the `Queries` namespace.

Remove the `FilterValidatableEndpoint` spec as this is in `/services/concerns`
when it is a controller concern and we don't explicitly test other controller
concerns; they are tested via the request specs and shared contexts. Update
shared context to cover the tests we are removing.
