--- app/services/concerns/filter_ignorable.rb
@@ -1,7 +1,0 @@
-module FilterIgnorable
-  extend ActiveSupport::Concern
-
-  def ignore?(filter:)
-    filter == :ignore || (!filter.nil? && filter.blank? && filter != false)
-  end
-end
