+++ app/services/concerns/queries/condition_formats.rb
