+++ app/services/concerns/queries/filter_ignorable.rb
@@ -1,0 +1,9 @@
+module Queries
+  module FilterIgnorable
+    extend ActiveSupport::Concern
+
+    def ignore?(filter:)
+      filter == :ignore || (!filter.nil? && filter.blank? && filter != false)
+    end
+  end
+end
