+++ app/services/concerns/queries/range_queries.rb
