+++ spec/services/concerns/queries/condition_formats_spec.rb
