+++ spec/services/concerns/queries/range_queries_spec.rb
